agent/requirements.txt
fastapi
uvicorn
langchain
langchain_google_genai
python-dotenv